+++ src/stuff.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: stuff.h,v 1.8 2000/08/10 09:13:47 grubba Exp $
+ * $Id: stuff.h,v 1.9 2000/08/10 09:20:23 grubba Exp $
  */
 #ifndef STUFF_H
 #define STUFF_H
@@ -13,7 +13,7 @@
 #include "global.h"
 
 /* Prototypes begin here */
-ptrdiff_t my_log2(size_t x);
+int my_log2(size_t x);
 int count_bits(unsigned INT32 x);
 int is_more_than_one_bit(unsigned INT32 x);
 double my_strtod(char *nptr, char **endptr);
